% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{AET.rast.max}
\alias{AET.rast.max}
\title{Raster method for Maximum monthly actual evapotranspiration.}
\usage{
AET.rast.max(block, jan.a = "a01", nmonth = 3)
}
\arguments{
\item{block}{Multi layer raster containing monthly climate statistics.}

\item{jan.a}{Name of first monthly actual evapotranspiration layer.}

\item{nmonth}{Number of months to assess peak actual evapotranspiration.}
}
\description{
Raster method for Maximum monthly actual evapotranspiration.
}
