% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{AET.rast.monthly}
\alias{AET.rast.monthly}
\title{Raster method for actual evapotranspiration (monthly)}
\usage{
AET.rast.monthly(block, jan.p = "p01", jan.e = "e01")
}
\arguments{
\item{block}{Multi layer raster containing monthly climate statistics.}

\item{jan.p}{Name of first monthly precipitation layer.}

\item{jan.e}{Name of first monthly evapotranspiration layer.}
}
\value{
Multi layer raster of actual evapotranspiration.
}
\description{
Calculates the maximum possible evapotranspiration based only on precipitation for each month, without monthly carry over of soil water storage.
}
