% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{AmplifiedReduction}
\alias{AmplifiedReduction}
\title{Reduction in horizontal resolution preserving vertical range.}
\usage{
AmplifiedReduction(hires, fact = 5)
}
\arguments{
\item{hires}{Input high resolution raster.}

\item{fact}{Aggregation factor to reduce resolution by.}
}
\value{
Reduced resolution raster maintaining the vertical range of the parent raster.
}
\description{
Reduction in horizontal resolution preserving vertical range.
}
