% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{ApplyClim.rast}
\alias{ApplyClim.rast}
\title{Apply a function across raster layers.}
\usage{
ApplyClim.rast(
  block,
  jan = "p01",
  mons = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12),
  fun = "sum",
  name = NULL
)
}
\arguments{
\item{block}{Multi layer raster containing monthly climate statistics.}

\item{jan}{Name of the first monthly climate statistic (assumes subsequent layers are same statistic)}

\item{mons}{Number of months to summarize.}

\item{fun}{Function to apply.}

\item{name}{Name to give to results.}
}
\value{
Raster summary of climatic statistic.
}
\description{
Apply a function across raster layers.
}
