% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateLong.R
\name{climateLong}
\alias{climateLong}
\title{Move monthly columns to single column for each parameter}
\usage{
climateLong(
  x,
  ID = "ID",
  name = "name",
  group = "group",
  lat = "lat",
  lon = "lon",
  elev = "elev",
  year = "year",
  p.select = "p",
  t.select = "t",
  th.select = "th",
  tl.select = "tl"
)
}
\arguments{
\item{x}{data frame containing relatively standard monthly climate data, ideally after selecting only one station of interest.}

\item{ID}{give the name of the column containing station identifier}

\item{name}{give the name of the column containing name}

\item{group}{give the name of the column containing a group membership}

\item{lat}{give the name of the column with latitude}

\item{lon}{give the name of the column with longitude}

\item{elev}{give the name of the column with elevation}

\item{year}{give the name of the column with year (optional if for multi year time series)}

\item{p.select}{prefix indicating precipitation columns (each monthly column identified by two digits after this prefix)}

\item{t.select}{prefix indicating mean temperature columns (each monthly column identified by two digits after this prefix)}

\item{th.select}{prefix indicating high temperature columns (each monthly column identified by two digits after this prefix)}

\item{tl.select}{prefix indicating low temperature columns (each monthly column identified by two digits after this prefix)}
}
\value{
Data frame with standardize names for columns and each month in a separate row.
}
\description{
Move monthly columns to single column for each parameter
}
\examples{
df <- climatools::Norms2010
selected <- subset(df, Station_ID \%in\% 'USW00094860')
ntab <- climateLong(selected, ID="Station_ID", name = 'Station_Name', group="State",lat = "Latitude",lon = "Longitude", elev = "Elevation", p.select = 'pp', year='Year_')
head(ntab)

}
