% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{comphillpos}
\alias{comphillpos}
\title{Compound hillslope position using 3 neighborhood scales}
\usage{
comphillpos(dm, r1, r2, r3)
}
\arguments{
\item{dm}{original raster}

\item{r1}{small neighborhood radius}

\item{r2}{intermediate neighborhood radius}

\item{r3}{large neighborhood radius}
}
\value{
Compound hillslope position raster
This function gives more weight to neighborhoods with higher relative relief.
}
\description{
Compound hillslope position using 3 neighborhood scales
}
\examples{
data("denali")
dem <- denali
dem <- toraster(dem)
dem <- reproject(dem, rs=250)
plot(dem)
x <- climatools::comphillpos(dem, 500, 2000, 5000)
plot(x)
}
