% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{conStats.rast}
\alias{conStats.rast}
\title{Constrained climate statistics}
\usage{
conStats.rast(
  r,
  first = "t01",
  nmon = c("semester", "quarter", 1),
  bot = 0,
  top = 100,
  fun = c("max", "min")
)
}
\arguments{
\item{r}{raster block containing at 12 consecutive months of temperature or other climate statistic}

\item{first}{Name of initial monthly value (January) for climate statistic.}

\item{nmon}{Number of months to compare means or alternatively quarters or semesters (with a fixed set of months for 3 or 6 month season, faster to calculate than a rolling through 12 months). When 12 is selected, then it is just an annual mean. If 1 month is selected, then a simple maximum or minimum of all 12 months can be calculated.}

\item{bot}{Bottom value to substitute for any values less than.}

\item{top}{Top value to substitute for any values greater than.}

\item{fun}{Function to rate consecutive means (either "max" or "min")}
}
\value{
Raster of a maximum or minimum climate statistic for a set number of months
}
\description{
This function calculates the maximum or minimum consecutive mean for a constrained climate statistic. For example, you could calculate the warmest 3 consecutive monthly temperature  constrained to between 0 and 30 degrees.
}
