% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{enhanceRast}
\alias{enhanceRast}
\title{Enhance raster using trends with a higher resolution digital elevation model.}
\usage{
enhanceRast(t1, e1, e2)
}
\arguments{
\item{t1}{Climate or other types of grids that should have a tight relationship with elevation.}

\item{e1}{Elevation grid matching resolution and extent of climate grid of interest.}

\item{e2}{Elevation grid of higher resolution, cropped to an area of interest.}
}
\value{
High resolution climate grid.
}
\description{
Enhance raster using trends with a higher resolution digital elevation model.
}
