% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationtools.R
\name{find.multithreshold}
\alias{find.multithreshold}
\title{Find Threshold for multiple Variables separating two classes.}
\usage{
find.multithreshold(x, class1, class2, variables)
}
\arguments{
\item{x}{Data frame of environmental and classification data.}

\item{class1}{Name of column with probabilities for class 1.}

\item{class2}{Name of column with probabilities for class 2.}

\item{variables}{Vector of column names with variables of interest.}
}
\value{
Reports a data frame of variable thresholds and variable strength in separating two classes.
}
\description{
Find Threshold for multiple Variables separating two classes.
}
