% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationtools.R
\name{find.threshold}
\alias{find.threshold}
\title{Find Threshold for Variable separating two classes.}
\usage{
find.threshold(x, class1, class2, variable)
}
\arguments{
\item{x}{Data frame of environmental and classification data.}

\item{class1}{Name of column with probabilities for class 1.}

\item{class2}{Name of column with probabilities for class 2.}

\item{variable}{Name of the column with variable of interest.}
}
\value{
Returns a list object identifying the optimal threshold of the variable to separate the two classes.
}
\description{
Find Threshold for Variable separating two classes.
}
