% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{focalmin}
\alias{focalmin}
\title{Get focal minimum with specified radius}
\usage{
focalmin(x, r)
}
\arguments{
\item{x}{raster}

\item{r}{radius}
}
\value{
focal minimum raster
This function makes use of lower resolution aggregated rasters to speed up calculations at increasing radii, compromising accuracy.
}
\description{
Get focal minimum with specified radius
}
