% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{fromraster}
\alias{fromraster}
\title{From raster to list object}
\usage{
fromraster(x)
}
\arguments{
\item{x}{Terra raster object.}
}
\value{
List object suitable for saving as package data.
}
\description{
From raster to list object
}
