% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTemp.R
\name{generatePpt}
\alias{generatePpt}
\title{Generate precipitation monthly time series}
\usage{
generatePpt(p.mean, p.max, p.min, m.max = 7, m.min = 1)
}
\arguments{
\item{p.mean}{mean monthly precipitation}

\item{p.max}{highest monthly precipitation}

\item{p.min}{lowest monthly precipitation}

\item{m.max}{number of month with highest precipitation}

\item{m.min}{number of month with lowest precipitation}
}
\value{
Monthly time series of precipitation for 12 month period.
This function is useful to explore artificial climatic scenarios. Precipitation follows a simple sine wave pattern if highest and lowest months are 6 months apart, and the mean precipitation is halfway between the highest and lowest values. Patterns other than a sine way will occur when high and low values are extreme or deviate from being 6 months apart. Only targeted mean monthly precipitation remains true in the latter case.
}
\description{
Generate precipitation monthly time series
}
\examples{
generatePpt(p.mean=100, p.max=120, p.min=80, m.max=8, m.min=2)
}
