% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTemp.R
\name{generateTemp}
\alias{generateTemp}
\title{Generate mean temperature monthly time series}
\usage{
generateTemp(t01, t07)
}
\arguments{
\item{t01}{January temperature (degrees Celsius)}

\item{t07}{July temperature (degrees Celsius)}
}
\value{
Monthly time series of mean temperatures for 12 month period.
This function is useful to explore artificial climatic scenarios. Temperatures follow a simple sine wave with January and July as the coldest and warmest months (or the inverse for the southern hemisphere). A real climate may show some slight asymmetry between warm months and cold months, and be shifted in phase by a month or more, but this curve is very close to the most frequent pattern.
}
\description{
Generate mean temperature monthly time series
}
\examples{
generateTemp(-5,21)
}
