% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{genrast}
\alias{genrast}
\title{Generate overlapping dummy rasters.}
\usage{
genrast(e)
}
\arguments{
\item{e}{4 value raster showing which raster is to exceed value of overlap extent.}
}
\value{
List of two rasters that overlap each other.
}
\description{
Function is used exclusively for demonstrating splice function examples.
}
