% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetDayLength}
\alias{GetDayLength}
\title{Get day length}
\usage{
GetDayLength(DayNumber, Lat)
}
\arguments{
\item{DayNumber}{Day of the year (1-365)}

\item{Lat}{Latitude (-90 - 90)}
}
\value{
Number of hours of daylight in a day.
}
\description{
Get day length
}
\examples{
GetDayLength(GetDayNumber(6),43)
}
