% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetDayNumber}
\alias{GetDayNumber}
\title{Get mean day of the year number by month}
\usage{
GetDayNumber(mon)
}
\arguments{
\item{mon}{Month of the year number (1-12).}
}
\value{
Day of the year at mid month (16-350).
}
\description{
Get mean day of the year number by month
}
\examples{
GetDayNumber(6)
}
