% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetDcl}
\alias{GetDcl}
\title{Get declination}
\usage{
GetDcl(DayNumber)
}
\arguments{
\item{DayNumber}{Day of the year}
}
\value{
Angle of declination of the earth relative to the sun for that time of year.
}
\description{
Get declination
}
\examples{
GetDcl(GetDayNumber(6))
}
