% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetNetSolar}
\alias{GetNetSolar}
\title{Get Net Solar Radiation}
\usage{
GetNetSolar(Ra, Elev, th, tl, p)
}
\arguments{
\item{Ra}{Calculated daily solar radiation hitting the earth's atmosphere}

\item{Elev}{Elevation above sea level (m).}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}

\item{p}{Mean monthly precipitation (mm)}
}
\value{
Net Daily Solar Radiation (\eqn{MJ m^{-2} day^{-1}}) based on formulas used in Walter et al (2000).
}
\description{
Get Net Solar Radiation
}
\examples{
GetNetSolar(GetSolarRad(GetDayNumber(6), 43), 210, 25, 15, 100)
}
\references{
Walter, I.A., Allen, R.G., Elliott, R., Jensen, M.E., Itenfisu, D., Mecham, B., Howell, T.A., Snyder, R., Brown, P., Echings, S. and Spofford, T., 2000. ASCE's standardized reference evapotranspiration equation. In Watershed management and operations management 2000 (pp. 1-11).
}
