% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetNumberDay}
\alias{GetNumberDay}
\title{Get number of days in month.}
\usage{
GetNumberDay(mon)
}
\arguments{
\item{mon}{Month of the year number (1-12).}
}
\value{
Numbers of days in the month (for average year, 28.25-31).
}
\description{
Get number of days in month.
}
\examples{
GetNumberDay(6)
}
