% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{GetPET.block}
\alias{GetPET.block}
\title{Raster method for mean monthly potential evapotranspiration (block of 12 months).}
\usage{
GetPET.block(mon, block, th.jan = "th01", tl.jan = "tl01", p.jan = "p01")
}
\arguments{
\item{mon}{Month number (1-12)}

\item{block}{Multi layer raster containing monthly high and low temperatures and preciptitation.}

\item{th.jan}{Name of layer representing first month of high temperatures (assuming the remaining 11 months follow in sequence)}

\item{tl.jan}{Name of layer representing first month of low temperatures (assuming the remaining 11 months follow in sequence)}

\item{p.jan}{Name of layer representing first month of precipitations (assuming the remaining 11 months follow in sequence)}
}
\value{
Raster of monthly mean potential evapotranspiration.
}
\description{
Raster method for mean monthly potential evapotranspiration (block of 12 months).
}
