% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{GetPET.rast}
\alias{GetPET.rast}
\title{Raster method for mean monthly potential evapotranspiration (1 month)}
\usage{
GetPET.rast(mon, th, tl, p)
}
\arguments{
\item{mon}{Month number (1-12)}

\item{th}{Mean daily high temperature raster (C).}

\item{tl}{Mean daily low temperature raster (C).}

\item{p}{Mean Monthly precipitation raster (mm).}
}
\value{
Raster of monthly mean potential evapotranspiration.
}
\description{
Raster method for mean monthly potential evapotranspiration (1 month)
}
