% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPET}
\alias{GetPET}
\title{Estimate potential evapotranspiration}
\usage{
GetPET(Ra, th, tl, p)
}
\arguments{
\item{Ra}{Calculated daily solar radiation hitting the earth's atmosphere}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}

\item{p}{Mean monthly precipitation (mm)}
}
\value{
Daily potential evapotranspiration (mm).
Adapted from formulas used in Walter et al (2000).
}
\description{
Estimate potential evapotranspiration
}
\references{
Walter, I.A., Allen, R.G., Elliott, R., Jensen, M.E., Itenfisu, D., Mecham, B., Howell, T.A., Snyder, R., Brown, P., Echings, S. and Spofford, T., 2000. ASCE's standardized reference evapotranspiration equation. In Watershed management and operations management 2000 (pp. 1-11).
}
