% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPETgs}
\alias{GetPETgs}
\title{Growing season adjusted potential evapotranspiration}
\usage{
GetPETgs(Ra, th, tl, p)
}
\arguments{
\item{Ra}{Calculated daily solar radiation hitting the earth's atmosphere}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}

\item{p}{Mean monthly precipitation (mm)}
}
\value{
Potential evapotranspiration (mm) adjusted for growing season.
This is the product of GetTransGrow() and GetPET() functions multiplied by a "crop" coefficient of 0.85 yields average results of similar magnitude within the United States as the older Thornthwaite method where it (the Thornthwaite method) was originally calibrated.
}
\description{
Growing season adjusted potential evapotranspiration
}
