% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPETthorn}
\alias{GetPETthorn}
\title{Thornthwaite 12 monthly potential evapotranspiration}
\usage{
GetPETthorn(mon, lat, t)
}
\arguments{
\item{mon}{Month of the year number (1-12).}

\item{t}{Mean daily mean temperature (degrees Celsius). Requires exactly 12 individual monthly mean temperatures to work properly.}

\item{Lat}{Latitude (-90 - 90)}
}
\value{
Monthly potential evapotranspiration (mm) from Thornthwaite as used in Lu et al (2005).
}
\description{
Thornthwaite 12 monthly potential evapotranspiration
}
\examples{
mon=c(1:12)
lat=43
t=generateTemp(15,25)
GetPETthorn(mon,lat,t)
}
