% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPETthorn2}
\alias{GetPETthorn2}
\title{Thornthwaite monthly potential evapotranspiration}
\usage{
GetPETthorn2(Dl, I, t)
}
\arguments{
\item{Dl}{Hours of day light per day.}

\item{I}{Parameter derived from 12 monthly temperatures via the getThornFactor() function}

\item{t}{Mean daily mean temperature (degrees Celsius)}
}
\value{
Thornthwaite monthly potential evapotranspiration (mm)
}
\description{
Thornthwaite monthly potential evapotranspiration
}
