% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrelief.R
\name{getrelief}
\alias{getrelief}
\title{Get topographic relief}
\usage{
getrelief(
  dm,
  r1,
  r2,
  n = 0,
  s = 0.1,
  p = c("low", "medium", "high", "exact"),
  breaks = NA
)
}
\arguments{
\item{dm}{Digital elevation model (terra raster) with projected coordinates.}

\item{r1}{Minimum focal radius for estimating relief (only focal radius when n-0).}

\item{r2}{Maximum focal radius for estimating relief.}

\item{n}{Number of segments between maximum and minimum radii to estimate relief.}

\item{s}{Slope for which relief is estimated (Default is 0.1 for a 10\% slope). Should be a ratio of vertical units to horizontal units).}

\item{p}{Precision from low to exact with higher levels of precision requiring more processing time.}

\item{breaks}{Optional vector of relief thresholds to specifically calculate a radius rather than to depend on interpolation among a regular interval between maximum and minimum radii.}
}
\value{
Topographic relief of either a fixed radius or for a fixed slope for a range of radii.
}
\description{
Get topographic relief
}
\examples{
library(terra)
#create fake raster
x <- (1:10)/10-0.5
y <- x
df <- merge(x,y)
df$z <- 1000/((df$x-0.25)^2+(df$y-0.40)^2+1)^2
set.seed(1)
df$z <- df$z +rnorm(10*10, sd = 100)
df$x <- df$x*10 -85
df$y <- df$y*10 +20
dem <- rast(cbind(x=df$x,y=df$y,z=df$z), type="xyz", crs='EPSG:4326')
dem <- reproject(dem=dem, rs = 250)
plot(dem)
#get relief with a 10\% slope using 10 intervals between focal radii and default low precision.
rng <- getrelief(dem, r1=1000, r2=100000, s=0.1, n=10)
#get relief with a 25\% slope using 10 intervals between focal radii and default low precision.
rng <- getrelief(dem, r1=1000, r2=100000, s=0.25, n=10)
plot(rng)
#get 10\% relief using specific break points of vertical relief and medium precision.
rng <- getrelief(dem, r1=1000, r2=100000, s=0.1, n=1, p='medium', breaks = c(300, 1000, 2500, 5000))
plot(rng)
#get relief with a fixed radius of 2000 meters.
rng <- getrelief(dem, r1=2000, s=0.1, n=0)
plot(rng)
}
