% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetSolar}
\alias{GetSolar}
\title{Estimate solar radiation reaching the earth's surface}
\usage{
GetSolar(Ra, Elev, th, tl, p)
}
\arguments{
\item{Ra}{Calculated daily solar radiation hitting the earth's atmosphere}

\item{Elev}{Elevation above sea level (m).}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}

\item{p}{Mean monthly precipitation (mm)}
}
\value{
Estimated solar radiation reaching the earth's surface. Based on a linear regression of 10 minute WorldClim 2.0 solar radiation estimates.
}
\description{
Estimate solar radiation reaching the earth's surface
}
