% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{GetSolarRad.rast}
\alias{GetSolarRad.rast}
\title{Raster method}
\usage{
GetSolarRad.rast(DayNumber, Raster)
}
\arguments{
\item{DayNumber}{Sequential date in the year (0-365).}

\item{Raster}{Any raster representing the target resolution and extent.}
}
\value{
Raster of solar radiation
}
\description{
Raster method
}
