% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetSolarRad}
\alias{GetSolarRad}
\title{Get solar radiation reaching the top of earth's atmosphere}
\usage{
GetSolarRad(DayNumber, Lat)
}
\arguments{
\item{DayNumber}{Day of the year (1-365)}

\item{Lat}{Latitude (-90 - 90)}
}
\value{
Daily solar radiation reaching the top of earth's atmosphere based on a formula used in Walter et al 2000.
}
\description{
Get solar radiation reaching the top of earth's atmosphere
}
\references{
Walter, I.A., Allen, R.G., Elliott, R., Jensen, M.E., Itenfisu, D., Mecham, B., Howell, T.A., Snyder, R., Brown, P., Echings, S. and Spofford, T., 2000. ASCE's standardized reference evapotranspiration equation. In Watershed management and operations management 2000 (pp. 1-11).
}
