% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{getThornFactor}
\alias{getThornFactor}
\title{Monthly temperature derived 'I' parameter for the Thornthwaite equation}
\usage{
getThornFactor(t)
}
\arguments{
\item{t}{Mean daily mean temperature (degrees Celsius). Requires exactly 12 individual monthly mean temperatures to work properly.}
}
\value{
Monthly temperature derived parameter for the parameter for the Thornthwaite equation.
}
\description{
Monthly temperature derived 'I' parameter for the Thornthwaite equation
}
