% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetTransGrow}
\alias{GetTransGrow}
\title{Estimate plant transpiration component of potential evapotranspiration mediated by likely safe growing season temperatures}
\usage{
GetTransGrow(th, tl)
}
\arguments{
\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}
}
\value{
Proportion of full potential evapotranspiration. Basically reduces the base formula for potential evapotranspiration to reflect plants may shut down their transpiration due to cold, leaving only evaporation from soil surfaces. This amount here is intended to be generalization for a diverse range of land cover, and is assumed 80% transpiration during growing season, and reduced to 0% during freezing conditions. Its actual amount is difficult to know, and would vary depending on plant cover.
}
\description{
Estimate plant transpiration component of potential evapotranspiration mediated by likely safe growing season temperatures
}
\examples{
GetTransGrow(25, 15)
GetTransGrow(10, -5)
GetTransGrow(1, -15)
}
