% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetVp}
\alias{GetVp}
\title{Estimate mean vapor pressure}
\usage{
GetVp(p, th, tl)
}
\arguments{
\item{p}{Mean monthly precipitation (mm)}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}
}
\value{
Mean vapor pressure for the day or month
Based on a linear regression using 10 minute WorldClim 2.0 (Fick and Hijmans, 2017) data vapor pressure estimates.
}
\description{
Estimate mean vapor pressure
}
\examples{
GetVp(100,25,15)
}
\references{
Fick, S.E. and Hijmans, R.J., 2017. WorldClim 2: new 1‐km spatial resolution climate surfaces for global land areas. International journal of climatology, 37(12), pp.4302-4315.
}
