% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{GrowTemp.rast}
\alias{GrowTemp.rast}
\title{Raster method to calculate growing season temperature.}
\usage{
GrowTemp.rast(block, t.jan = "t01")
}
\arguments{
\item{block}{Multi layer raster containing mean monthly temperature for 12 months.}

\item{t.jan}{Name of first monthly temperature layer.}
}
\value{
Raster of positive temperature of the warmest 6 months.
}
\description{
Raster method to calculate growing season temperature.
}
