% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{GrowTemp}
\alias{GrowTemp}
\title{Growing season temperature}
\usage{
GrowTemp(t)
}
\arguments{
\item{t}{Vector of monthly mean temperatures for 12 months (degrees Celsius)}
}
\value{
Mean positive temperature of warmest 6 consecutive months (degrees Celsius; all monthly temperatures below zero counted as zero)
}
\description{
Growing season temperature
}
\examples{
t <- generateTemp(-5,21)
GrowTemp(t)
}
