% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{hillpos}
\alias{hillpos}
\title{Calculate hillslope position}
\usage{
hillpos(dm, xmax, xmin, xmed)
}
\arguments{
\item{dm}{original raster}

\item{xmax}{raster with focal neighborhood maximum}

\item{xmin}{raster with focal neighborhood minimum}

\item{xmed}{raster with focal neighborhood median}
}
\value{
Raster with relative slope position 0 to 1.
}
\description{
Calculate hillslope position
}
