% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{hillshade}
\alias{hillshade}
\title{Create hillshade directly from DEM raster}
\usage{
hillshade(x, angle = 45, direction = 0)
}
\arguments{
\item{x}{Digital elevation model}

\item{angle}{The elevation angle(s) of the light source (sun), in degrees}

\item{direction}{The direction (azimuth) angle(s) of the light source (sun), in degrees}
}
\value{
Hillshade raster for use to visualize terrain texture.
This function combines terra::terrain() function to create slope and aspect rasters, then employs the terra::shade() function to generate hillshade in one step.
}
\description{
Create hillshade directly from DEM raster
}
