% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClimplot.R
\name{makeClimplot}
\alias{makeClimplot}
\title{Make a monthly climate graph}
\usage{
makeClimplot(
  t,
  p,
  th = NULL,
  tl = NULL,
  e = NULL,
  a = NULL,
  u = NULL,
  t80 = NULL,
  t20 = NULL,
  p80 = NULL,
  p20 = NULL,
  name = NULL,
  bar = FALSE
)
}
\arguments{
\item{t}{mean monthly temperature (degrees Celsius)}

\item{p}{mean monthly preciptation (mm)}

\item{th}{mean daily high temperature (degrees Celsius)}

\item{tl}{mean daily low temperature (degrees Celsius)}

\item{e}{potential evapotranspiration (mm) [optional]}

\item{a}{actual evapotranspiration (mm) [optional]}

\item{u}{surplus precipitation (mm) [optional]}

\item{t80}{80th percentile temperature  [optional](not yet implemented)}

\item{t20}{20th percentile temperature  [optional](not yet implemented)}

\item{p80}{80th percentile precipitation  [optional](not yet implemented)}

\item{p20}{20th percentile precipitation  [optional](not yet implemented)}

\item{name}{name of the climate station}

\item{bar}{TRUE/FALSE option to show bar graph (TRUE) or a smoothed curve.}
}
\value{
Climate graph
}
\description{
Make a monthly climate graph
}
\examples{
selected = subset(climatools::Norms2010, Station_ID \%in\% 'USW00023234')
climtab <- climateLong(selected, name <- 'Station_Name',lat = "Latitude",lon = "Longitude", elev = "Elevation", p.select = 'pp', year='Year_')
climtab <- climtab |> group_by(name, lat, lon, elev, mon) |> summarise(across(c("t","th","tl","p"),mean))
climtab <- climtab |> mutate(e = GetPET(mon, lat, th, tl, p))
budget = SoilWaterBudget(climtab$p, climtab$e, 150)
climtab <- climtab |> mutate(s=budget$s,a=budget$a, u=budget$u, d=budget$d)

makeClimplot(t=climtab$t,
th=climtab$th,
tl=climtab$tl,
p=climtab$p,
e=climtab$e,
a=climtab$a,
u=climtab$u,
name=climtab$name,
bar=F)

makeClimplot(t=climtab$t,
th=climtab$th,
tl=climtab$tl,
p=climtab$p,
e=climtab$e,

name=climtab$name,
bar=F)

makeClimplot(t=climtab$t,
th=climtab$th,
tl=climtab$tl,
p=climtab$p,

name=climtab$name,
bar=F)

makeClimplot(t=climtab$t,
th=climtab$th,
tl=climtab$tl,
p=climtab$p,
e=climtab$e,

name=climtab$name,
bar=T)

}
