% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{makexyrast}
\alias{makexyrast}
\title{Make XY Raster}
\usage{
makexyrast(x, rotations = 0)
}
\arguments{
\item{x}{raster to extract xy coordinates}

\item{rotations}{Specify number of rotations of xy coordinates (default zero)}
}
\value{
Multi channel raster with xy coordinates. (rotations add alternative angles a random forest covariates). Rasters named "lat" for y coordinate, and "lon"  for x coordinate.
}
\description{
Make XY Raster
}
\examples{
x <- terra::rast(matrix(1:25, nrow=5, ncol=5))
xyrast <- makexyrast(x)
}
