% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationtools.R
\name{maxKappa}
\alias{maxKappa}
\title{maxKappa}
\usage{
maxKappa(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual presence/absences, 0 or 1.}

\item{predicted}{Vector of predicted presence between 0 and 1.}
}
\value{
Script loops through classification thresholds between 0 and 1, and reports the kappa statistic for the threshold with the most information.
}
\description{
maxKappa
}
