% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{refitrast}
\alias{refitrast}
\title{Re-fit climate later to DEM}
\usage{
refitrast(
  x,
  elev = NULL,
  cropto = NULL,
  cropfrom = NULL,
  rotations = 0,
  sampdens = 1500,
  altlayer = NULL
)
}
\arguments{
\item{x}{climatic layer to re-fit.}

\item{elev}{optional digital elevation model of similar resolution as climatic layer.}

\item{cropto}{spatial extent of cropped area (xmin, xmax, ymin, ymax)}

\item{cropfrom}{spatial extent of problematic area to be removed from model (xmin, xmax, ymin, ymax)}

\item{rotations}{Optional number of rotation of XY coordinates used for smoother random forest model of linear model residuals.}

\item{sampdens}{sample density for extracting points to train models.}

\item{altlayer}{optional alternative covariate layer(s) (can be multiple rasters combined, e.g. water body influence layer, aspects)}
}
\value{
Re-fitted climatic model.
}
\description{
This function can be used to omit problematic portions of a climatic layer and refit to a digital elevation model. The results can be used to patch problems in larger climatic layers. Omitting elevation layer bases refit exclusively on trends in xy coordinates.
}
\examples{
library(climatools)
library(terra)
#load in raw data to create temperature raster
climatedata <- read.csv(system.file("extdata", "pts.t.csv", package="climatools"))
altlayer <- rast(system.file("extdata", "altlayer.tif", package="climatools") )
#July temperature and include station elevation
pts <- data.frame(x=climatedata$lon, y=climatedata$lat, z=climatedata$t07, elev=climatedata$elev)
#crop to northern Great Plains
cropto <- c(-110,-88, 40, 50)

#How the layer should look
t07 <- toclimrast(pts, altlayer[[1:2]], cropto, covrange = 500, randforest = F)
plot(t07, col=map.pal('bcyr'))

#Artificially move a data point too far north, creating an anomaly.
pts[1342,]$y <- pts[1342,]$y+14
points(vect(pts[1342,], geom=c("x", "y"),crs=crs('epsg:4326')))
t07a <- toclimrast(pts, altlayer[[1:2]], cropto, covrange = 500, randforest = F)
plot(t07a, col=map.pal('bcyr'))
points(vect(pts[1342,], geom=c("x", "y"),crs=crs('epsg:4326')))

#remove the anomaly
cropto <- c(-105,-95,43,49)
cropout <- c(-102,-97,45,48)
t07aout <- refitrast(t07a, altlayer$elev, cropto, cropout, rotations=3)

#Splice to patch anomalous layer
t07spliced <- spliceraster(t07aout, t07a)
plot(t07spliced, col=map.pal('bcyr'))
points(vect(pts[1342,], geom=c("x", "y"),crs=crs('epsg:4326')))
}
