% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject.R
\name{reproject}
\alias{reproject}
\title{Reproject any raster to Lambert_Azimuthal_Equal_Area with specified center and extent.}
\usage{
reproject(
  dem,
  lat = NA,
  lon = NA,
  rs = NA,
  h = NA,
  w = NA,
  prj = NA,
  method = NA
)
}
\arguments{
\item{dem}{Digital elevation model or other type of raster}

\item{lat}{center latitude (default is existing raster)}

\item{lon}{center longitude (default is existing raster)}

\item{rs}{resolution of raster in meters (default is converted from existing raster)}

\item{h}{extent north to south in meters (default is converted from existing raster)}

\item{w}{extent east to west in meters (default is converted from existing raster)}

\item{prj}{Coordinate reference system (crs); Defaults to equal area azimuth.}

\item{method}{Resampling method: "near", "bilinear", "cubic", "cubicspline","lanczos", "min", "max", etc. See terra::project documentation.}
}
\value{
new raster with Lambert_Azimuthal_Equal_Area projection
}
\description{
Reproject any raster to Lambert_Azimuthal_Equal_Area with specified center and extent.
}
