% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{RestoreMaxMin}
\alias{RestoreMaxMin}
\title{Restore maximum and minimum values of a resampled raster}
\usage{
RestoreMaxMin(x, y, s = NA, e = NA)
}
\arguments{
\item{x}{Original raster.}

\item{y}{Resampled or reprojected raster.}

\item{s}{Optional point simple feature vector (sf) with locations of known high points, to further enhance accuracy of elevation extremes that may have been smoothed out in DEM resampling.}

\item{e}{Optional name of the column corresponding to elevation in the simple feature vector.}
}
\value{
Resampled raster with original maximum and minimum extreme values restored.
This function is used to correct for the loss in highest and lowest values within a local neighborhood after resampling (i.e. bilinear method) blends them with adjacent values (or substitutes with nearest neighbor in the case of resampling method = "near", or overshoots in the case of "cubic" methods.).This tool is intended to preserve full range of values when this is deemed more important.
}
\description{
Restore maximum and minimum values of a resampled raster
}
