% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{rotatexy}
\alias{rotatexy}
\title{Rotate XY Coordinates}
\usage{
rotatexy(x, y, a, cx = NA, cy = NA)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}

\item{a}{angle in degrees}

\item{cx}{optional center of rotation x coordinate (default is center of point cloud)}

\item{cy}{optional center of rotation y coordinate (default is center of point cloud)}
}
\value{
data frame of rotated xy coordinates
}
\description{
Rotate XY Coordinates
}
\examples{
df <- data.frame(
x=runif(10,0,10),
y=rnorm(10,5,5))
df2 <-  rotatexy(df$x,df$y, a=2)
plot(df$y ~ df$x)
points(df2$y ~ df2$x, col='red')
}
