% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{rposition}
\alias{rposition}
\title{Raster positions relative to each other}
\usage{
rposition(r1, r2)
}
\arguments{
\item{r1}{First raster}

\item{r2}{Second raster}
}
\value{
Vector of four values corresponding to xmin, xmax, ymin, and ymax of the spatial extent of the area of overlap, wherein a 1 corresponding to where first raster exceeds the extent, while a -1 corresponds to raster 2 exceeding the extent. A 0 values means that both rasters share the same extent value.
}
\description{
The function assesses the relative positions of two overlapping rasters expressed as a 4 value vector.
}
