% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject.R
\name{setProjection}
\alias{setProjection}
\title{Set a custom projection.}
\usage{
setProjection(
  prj = c("equalarea.azimuthal", "equaldistant.azimuthal", "conformal.azimuthal",
    "equalarea.cylindrical", "equaldistant.cylindrical", "conformal.cylindrical",
    "equalarea.conic", "equaldistant.conic", "conformal.conic", "conformal.transverse",
    "conformal.oblique", "equalarea.transverse", "equalarea.oblique", "geographic"),
  datum = c("WGS84", "NAD83", "NAD27"),
  unit = c("m", "ft", "usft"),
  lat = NA,
  lon = NA,
  lat2 = NA,
  lon2 = NA,
  feast = 0,
  fnorth = 0,
  orglat = NA,
  scf = 0.9996
)
}
\arguments{
\item{prj}{Projection method among 12 options (default is equal area azimuthal).}

\item{datum}{Set of three common datums (default is WGS84).}

\item{unit}{Set units of projection (default is meters)}

\item{lat}{Latitude of center of projection, first parallel of a conic projection, or first point of an oblique Mercator projection.}

\item{lon}{Longitude of center of projection, center meridian of a conic projection (for scaling), or first point of an oblique Mercator projection.}

\item{lat2}{Latitude of second parallel of a conic projection (for scaling), or second point of an oblique Mercator projection (to establish scaling axis).}

\item{lon2}{Longitude of second point of an oblique Mercator projection (to establish scaling axis).}

\item{feast}{False easting.}

\item{fnorth}{False northing.}

\item{orglat}{Latitude of coordinate origins.}

\item{scf}{Scaling factor of Transverse and Oblique Mercator projections.}
}
\value{
Well Known Text projection string to be used in raster and vector coordinate reference definitions.
}
\description{
Set a custom projection.
}
