% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{shadowMax}
\alias{shadowMax}
\title{Directional Maximum}
\usage{
shadowMax(x, r, asp = 270, p = c("low", "medium", "high", "exact"))
}
\arguments{
\item{x}{raster}

\item{r}{radius (meters)}

\item{asp}{aspect or direction of the wind.}

\item{p}{precision, from low to exact, with higher levels of precision requiring more processing time}
}
\value{
Raster with directional maximum with effects tapering to zero at radius, spread across 45 degrees. Subtracting elevation from directional maximum results in the equivalent to a rain shadow in the direction opposite of the specified aspect or wind direction.
}
\description{
Directional Maximum
}
