% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilWaterBudget.R
\name{SoilWaterBudget}
\alias{SoilWaterBudget}
\title{Soil water budget}
\usage{
SoilWaterBudget(p, e, awc)
}
\arguments{
\item{p}{Monthly precipitation (mm)}

\item{e}{Monthly potential evapotranspiration (mm)}

\item{awc}{Soil available  water holding capacity (mm)}
}
\value{
Data frame with monthly time series of:\tabular{ll}{
   \code{a: } \tab actual evapotranspiration (mm) \cr
   \tab \cr
   \code{s: } \tab soil water (mm) \cr
   \tab \cr
   \code{u: } \tab precipitation surplus (mm) \cr
   \tab \cr
   \code{d: } \tab soil moisture deficit (mm) \cr
}
}
\description{
Soil water budget
}
