% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{spliceraster}
\alias{spliceraster}
\title{Splice Rasters}
\usage{
spliceraster(r1, r2)
}
\arguments{
\item{r1}{First raster}

\item{r2}{Second raster}
}
\value{
Raster combination of two rasters.
}
\description{
Function blends two terra rasters using smooth gradients. Alternative to mosaic which only averages the area of overlap.
}
\examples{
rg <- genrast(c(0,1,1,0))
r1 <- rg[[1]]
r2 <- rg[[2]]
plot(mosaic(r1,r2))
m1 <- smoothmosaic(r1,r2)
plot(m1)
}
