% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{SummerTemp}
\alias{SummerTemp}
\title{Warmest Trimester}
\usage{
SummerTemp(t)
}
\arguments{
\item{t}{Vector of monthly mean temperatures for 12 months (degrees Celsius)}
}
\value{
Mean positive temperature of warmest 3 consecutive months (degrees Celsius; all monthly temperatures below zero counted as zero)
}
\description{
Warmest Trimester
}
\examples{
t <- generateTemp(-5,21)
GrowTemp(t)
}
