% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{toraster}
\alias{toraster}
\title{To raster from list object}
\usage{
toraster(x)
}
\arguments{
\item{x}{List object imported from package data.}
}
\value{
Terra raster object.
}
\description{
To raster from list object
}
\examples{
data(denali)
denali <- denali
denali <- toraster(denali)
plot(denali)
}
