% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{tpi}
\alias{tpi}
\title{Topographic position index using 3 neighborhood scales}
\usage{
tpi(max1, max2, max3, min1, min2, min3, med1, med2, med3)
}
\arguments{
\item{max1}{small neighborhood focal maximum}

\item{max2}{intermediate neighborhood focal maximum}

\item{max3}{large neighborhood focal maximum}

\item{min1}{small neighborhood focal minimum}

\item{min2}{intermediate neighborhood focal minimum}

\item{min3}{large neighborhood focal minimum}

\item{med1}{small neighborhood focal median}

\item{med2}{intermediate neighborhood focal median}

\item{med3}{large neighborhood focal median}
}
\value{
Topographic position index raster
This function simply averages the hillslope positions for all 3 neighborhoods.
}
\description{
Topographic position index using 3 neighborhood scales
}
