% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{warm6ppt}
\alias{warm6ppt}
\title{Highest/Lowest n month period temperature/precipitation}
\usage{
warm6ppt(
  t,
  p,
  n = 6,
  max_var = TRUE,
  v_input = c("t", "p"),
  v_return = c("t", "p", "b"),
  bt = 0
)
}
\arguments{
\item{t}{vector of 12 monthly mean temperatures}

\item{p}{vector of 12 monthly mean precipitation}

\item{n}{number of months in period}

\item{max_var}{Use maximum if TRUE (minimum if FALSE)}

\item{v_input}{Assess which parameter to rate extreme for period (t = temperature; p = precipitation)}

\item{v_return}{Return which parameter for the extreme period if TRUE (t = temperature; p = precipitation; b = positive biotemperature)}

\item{bt}{Temperature threshold for biotemperature}
}
\value{
A single value for the temperature/precipitation of the warmest/coldest/wettest/driest n-month period.
}
\description{
Highest/Lowest n month period temperature/precipitation
}
\examples{
t=generateTemp(5,22)
p=generatePpt(100,50,150)
#precipitation of warmest 6 months
warm6ppt(t, p, n=6, max_var=TRUE, v_input='t', v_return='p')
#temperature of warmest 6 months
warm6ppt(t, p, n=6, max_var=TRUE, v_input='t', v_return='t')
#temperature of driest 3 months
warm6ppt(t, p, n=3, max_var=FALSE, v_input='p', v_return='t')
#precipitation of driest 3 months
warm6ppt(t, p, n=3, max_var=FALSE, v_input='p', v_return='p')
#temperature of wettest 3 months
warm6ppt(t, p, n=3, max_var=TRUE, v_input='p', v_return='t')
#positive biotemperature of warmest 6 months (default threshold of 0 degrees Celsius)
warm6ppt(t, p, n=6, max_var=TRUE, v_input='t', v_return='b')
#positive biotemperature of warmest 6 months, with a threshold of 5 degrees Celsius
warm6ppt(t, p, n=6, max_var=TRUE, v_input='t', v_return='b', bt = 5)
#positive biotemperature of 12 months
warm6ppt(t, p, n=12, max_var=TRUE, v_input='t', v_return='b')
#positive biotemperature of 12 months, with a threshold of 5 degrees Celsius
warm6ppt(t, p, n=12, max_var=TRUE, v_input='t', v_return='b', bt = 5)

}
