% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{XtremLow.rast}
\alias{XtremLow.rast}
\title{Raster method for extreme annual low temperature.}
\usage{
XtremLow.rast(Tcl, Tc = NULL, Tw = NULL)
}
\arguments{
\item{Tcl}{Raster of Mean daily low temperature for coldest month.}

\item{Tc}{Raster of Mean temperature for coldest month (optional).}

\item{Tw}{Raster of Mean temperature for warmest month (optional).}
}
\value{
Estimated mean annual extreme low temperature based on analysis other data sets.
}
\description{
Raster method for extreme annual low temperature.
}
