% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{XtremLow}
\alias{XtremLow}
\title{Estimate extreme annual low temperature}
\usage{
XtremLow(Tcl, Lat, Lon, Elev)
}
\arguments{
\item{Tcl}{Mean daily low temperature of coldest month (degrees Celsius)}

\item{Lat}{Latitude (-90 - 90)}

\item{Lon}{Longitude (-180 - 180)}

\item{Elev}{Elevation above sea level (m).}
}
\value{
Extreme annual low temperature (degrees Celsius) as used for plant hardiness zones.
Estimated with a linear regression against previously published hardiness zone maps (Magarey, Borchert, & Schlegel, 2008; Daly et al., 2012) and gridded mean daily low temperature layers (Hijmans and Cameron, 2005).
}
\description{
Estimate extreme annual low temperature
}
\examples{
XtremLow(15, 24, -82, 0)
}
\references{
Daly, C., Widrlechner, M.P., Halbleib, M.D., Smith, J.I. and Gibson, W.P., 2012. Development of a new USDA plant hardiness zone map for the United States. Journal of Applied Meteorology and Climatology, 51(2), pp.242-264.

Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978. (http://www.worldclim.org/version1)

Magarey, R.D., Borchert, D.M. and Schlegel, J.W., 2008. Global plant hardiness zones for phytosanitary risk analysis. Scientia Agricola, 65(SPE), pp.54-59.
}
