% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{Biotemperature}
\alias{Biotemperature}
\title{Mean annual biotemperature}
\usage{
Biotemperature(t)
}
\arguments{
\item{t}{Vector of monthly mean temperatures for 12 months (degrees Celsius)}
}
\value{
Mean annual positive temperature (degrees Celsius; all monthly temperatures below zero counted as zero then averaged)
}
\description{
Mean annual biotemperature
}
\examples{
t <- generateTemp(-5,21)
Biotemperature(t)
}
