% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{comphillpos}
\alias{comphillpos}
\title{Compound hillslope position using 3 neighborhood scales}
\usage{
comphillpos(dm, max1, max2, max3, min1, min2, min3, med1, med2, med3)
}
\arguments{
\item{dm}{original raster}

\item{max1}{small neighborhood focal maximum}

\item{max2}{intermediate neighborhood focal maximum}

\item{max3}{large neighborhood focal maximum}

\item{min1}{small neighborhood focal minimum}

\item{min2}{intermediate neighborhood focal minimum}

\item{min3}{large neighborhood focal minimum}

\item{med1}{small neighborhood focal median}

\item{med2}{intermediate neighborhood focal median}

\item{med3}{large neighborhood focal median}
}
\value{
Compound hillslope position raster
This function gives more weight to neighborhoods with higher relative relief.
}
\description{
Compound hillslope position using 3 neighborhood scales
}
