% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{enhanceRast}
\alias{enhanceRast}
\title{Enhance raster using trends with a higher resolution digital elevation model.}
\usage{
enhanceRast(t1, e1, e2)
}
\arguments{
\item{t1}{Climate or other types of grids that should have a tight relationship with elevation.}

\item{e1}{Elevation grid matching resolution and extent of climate grid of interest.}

\item{e2}{Elevation grid of higher resolution, cropped to an area of interest.}
}
\value{
High resolution climate grid.
}
\description{
Enhance raster using trends with a higher resolution digital elevation model.
}
\examples{
#load package data for elevation
data("denali")
dem0 <- denali
dem0 <- toraster(dem0)
dem <- reproject(dem0, rs=250)
plot(dem)

#load package data for July temperature
data("Tw")
temperature0 <- Tw
temperature0 <- toraster(temperature0)

#Original temperature is coarse resolution.
plot(temperature0)

#This resamples to match resolution of elevation, but it appears blurry.
temperature <- project(temperature0, dem)
plot(temperature)

#prepare lower resolution elevation raster matching temperature resolution
dem0 <- project(dem0, temperature0)
plot(dem0)
#This estimates a local relationship of temperature with elevation.
temperature <- enhanceRast(temperature0, dem0,  dem)
plot(temperature)
}
