% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{focalmax}
\alias{focalmax}
\title{Get focal maximum with specified radius}
\usage{
focalmax(x, r)
}
\arguments{
\item{x}{raster}

\item{r}{radius}
}
\value{
focal maximum raster
This function makes use of lower resolution aggregated rasters to speed up calculations at increasing radii, compromising accuracy.
}
\description{
Get focal maximum with specified radius
}
