% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{focalmed}
\alias{focalmed}
\title{Get focal median with specified radius}
\usage{
focalmed(x, r)
}
\arguments{
\item{x}{raster}

\item{r}{radius}
}
\value{
focal median raster
This function makes use of lower resolution aggregated rasters to speed up calculations at increasing radii, compromising accuracy.
}
\description{
Get focal median with specified radius
}
