% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPET}
\alias{GetPET}
\title{Monthly potential evapotranspiration}
\usage{
GetPET(mon, lat, th, tl, p)
}
\arguments{
\item{mon}{Month of the year number (1-12)}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}

\item{p}{Mean monthly precipitation (mm)}

\item{Lat}{Latitude (-90 - 90)}
}
\value{
Potential evapotranspiration (mm/month) adjusted for growing season.
This is the product of GetTransGrow() and GetPETDaily() functions multiplied by a "crop" coefficient of 0.85 and number of days per month, which yields average results of similar magnitude within the United States as the older Thornthwaite method where it (the Thornthwaite method) was originally calibrated.
}
\description{
Monthly potential evapotranspiration
}
\examples{
GetPET(mon = 6, lat = 43, th=25, tl=15, p=100)
}
