% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetPETthorn2}
\alias{GetPETthorn2}
\title{Thornthwaite monthly potential evapotranspiration}
\usage{
GetPETthorn2(Dl, I, t)
}
\arguments{
\item{Dl}{Hours of day light per day.}

\item{I}{Parameter derived from 12 monthly temperatures via the getThornFactor() function}

\item{t}{Mean daily mean temperature (degrees Celsius)}
}
\value{
Thornthwaite monthly potential evapotranspiration (mm)
}
\description{
Thornthwaite monthly potential evapotranspiration
}
\examples{
mon <- 6 #June
lat <- 43 #north temperate zone
Dl <- GetDayLength(GetDayNumber(mon), lat) #day length
t.12months=generateTemp(-5,21) #monthly temperatures for a whole 12 month period
I = getThornFactor(t.12months) #'I' parameter
t = t.12months[6]#June temperature
GetPETthorn2(Dl,I, t)
}
