% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrelief.R
\name{getrelief}
\alias{getrelief}
\title{Get topographic relief}
\usage{
getrelief(dm, r1, r2, n = 0, s = 0.1)
}
\arguments{
\item{dm}{Digital elevation model (terra raster) with projected coordinates.}

\item{r1}{Minimum focal radius for estimating relief (only focal radius when n-0).}

\item{r2}{Maximum focal radius for estimating relief.}

\item{n}{Number of segments between maximum and minimum radii to estimate relief.}

\item{s}{Slope for which relief is estimated (Default is 0.1 for a 10% slope). Should be a ratio of vertical units to horizontal units).}
}
\value{
Topographic relief of either a fixed radius or for a fixed slope for a range of radii.
}
\description{
Get topographic relief
}
