% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPET.R
\name{GetVp}
\alias{GetVp}
\title{Estimate mean vapor pressure}
\usage{
GetVp(p, th, tl)
}
\arguments{
\item{p}{Mean monthly precipitation (mm)}

\item{th}{Mean daily high temperature (degrees Celsius)}

\item{tl}{Mean daily low temperature (degrees Celsius)}
}
\value{
Mean vapor pressure for the day or month
Based on a linear regression using 10 minute WorldClim 2.0 data vapor pressure estimates.
}
\description{
Estimate mean vapor pressure
}
\examples{
GetVp(100,25,15)
}
