% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{hillpos}
\alias{hillpos}
\title{Calculate hillslope position}
\usage{
hillpos(dm, r)
}
\arguments{
\item{dm}{original raster}

\item{r}{neighborhood radius}
}
\value{
Raster with relative slope position 0 to 1.
}
\description{
Calculate hillslope position
}
\examples{
data("denali")
dem <- denali
dem <- toraster(dem)
dem <- reproject(dem, rs=250)
plot(dem)
x <- climatools::hillpos(dem, 2000)
plot(x)
}
