% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{refit}
\alias{refit}
\title{Re-fit climate later to DEM}
\usage{
refit(
  x,
  elev = NULL,
  cropto = NULL,
  cropfrom = NULL,
  rotations = 0,
  sampdens = 1500,
  altlayer = NULL
)
}
\arguments{
\item{x}{climatic layer to re-fit.}

\item{elev}{optional digital elevation model of similar resolution as climatic layer.}

\item{cropto}{spatial extent of cropped area (xmin, xmax, ymin, ymax)}

\item{cropfrom}{spatial extent of problematic area to be removed from model (xmin, xmax, ymin, ymax)}

\item{rotations}{Optional number of rotation of XY coordinates used for smoother random forest model of linear model residuals.}

\item{sampdens}{sample density for extracting points to train models.}

\item{altlayer}{optional alternative covariate layer(s) (can be multiple rasters combined, e.g. water body influence layer, aspects)}
}
\value{
Re-fitted climatic model.
}
\description{
This function can be used to omit problematic portions of a climatic layer and refit to a digital elevation model. The results can be used to patch problems in larger climatic layers. Omitting elevation layer bases refit exclusively on trends in xy coordinates.
}
