% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilWaterBudget.R
\name{SoilWaterBudget}
\alias{SoilWaterBudget}
\title{Soil water budget}
\usage{
SoilWaterBudget(p, e, awc)
}
\arguments{
\item{p}{Monthly precipitation (mm)}

\item{e}{Monthly potential evapotranspiration (mm)}

\item{awc}{Soil available  water holding capacity (mm)}
}
\value{
Dataframe with monthly time series of:
}
\description{
Soil water budget
}
