% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{toclimrast}
\alias{toclimrast}
\title{Make climate raster from point data}
\usage{
toclimrast(
  pts,
  altlayer,
  cropto = NULL,
  covrange = 0,
  minrow = 50,
  segx = 5,
  segy = 5,
  cropbuffer = 5,
  randforest = TRUE,
  smoothresiduals = TRUE,
  refit = FALSE
)
}
\arguments{
\item{pts}{Data frame of 3 or 4 columns; first two: xy coordinates (longitude and latitude decimal degrees); third column is climate attribute like temperature; fourth column if provided is substitute values for first covariate (e.g. station elevation being more precise than gridded elevation).}

\item{altlayer}{Terra raster layer (multiple layers) representing covariates. Elevation is recomended as first covariate if station elevation provided as fourth column in point data.}

\item{cropto}{Optional crop extent c(xmin, xmax, ymin, ymax).}

\item{covrange}{Minimum range in first covariate before building submodel coefficients.}

\item{minrow}{Minimum number of data points before building submodel coefficients.}

\item{segx}{Number of segments in x axis for building submodel coefficients.}

\item{segy}{Number of segments in y axis for building submodel coefficients.}

\item{cropbuffer}{Buffer around the crop extent to incorporate points into model outside cropped area (units of raster projection).}

\item{randforest}{Use random forest model if true (requires ranger package) to generate coefficients and residuals layers. Alternatively, coefficients and residuals layers are interpolated solely using xy coordinate using gstats package.}

\item{smoothresiduals}{Residuals by default (TRUE) are generated at a lower resolution and smoothed to remove blocky appearance. Can turn off (FALSE) for lower resolution rasters.}

\item{refit}{Alternative to smooth residuals by resampling residuals raster with another random forest model for a less blocky appearance (more time intensive).}
}
\value{
Climate raster matching cropped extent.
}
\description{
This function generates a climatic raster from point data. The user provides covariate rasters and a data frame consisting of xy coordinates and either temperature or precipitation data.   The lapse rate of temperature with elevation and the direction of rain shadows vary over large geographic extent. Therefore, this function first establishes local linear regression coefficients via a moving window. These coefficients are then rasterized and incorporated as interaction terms in a global linear model. Residuals are rendered (via random forest or interpolation) and smoothed and then added to the model in the final step.
A Linear model  was used to build the main model because it can extrapolate where local station data has limited elevation range, unlike random forest models. Random forest also cannot show gradients in flat areas very well unless it is smoothed, but smoothing would destroy sharp gradients in mountains. But as residuals, a random forest model can recover some of the nuance like local temperature inversions that are otherwise lost in a simple linear model. Alternatively, if interpolation is used for residuals, then only the relative spatial positions of station points is rendered without incorporating other covariates.
}
\details{
Zero limited precipitation and daily temperature range data should be transformed to avoid modeling negative values. Covariates are extracted from the raster to the points. As an option, a fourth variable can be added to the point data to allow for actual station elevation to be used in building the model instead of the low resolution raster. Raster output will reflect the diminished elevation range of the covariate raster (but this can be amplified using the enhanceRast function). Using actual elevation is probably less critical for precipitation as precipitation is mainly related to elevation in context of uplift caused by its topographic neighborhood.
}
\examples{
library(terra)
library(climatools)
#load multilayer raster of 1km or 4km resolution (can be in decimal degrees units). This layer includes elevation, proximity to water, and rain shadows for west and south winds.
altlayer <- rast(system.file("extdata", "altlayer.tif", package="climatools") )
 #include only layers relevent to model, like elevation, relative elevation, a focal mean water bodies with a neighborhood of 50 km, and SAGA GIS derived wind effects analysis to identify rain shadows for 8 wind directions.Ensure that elevation is listed first to constrain point data by the range in this variable.
 #load point observations of climate data (precipitation)
 climatedata <- read.csv(system.file("extdata", "pts.p.csv", package="climatools"))
 #create a 3 column data set identifying rainfall for July as climate attribute to model.
 pts <- data.frame(x=climatedata$lon, y=climatedata$lat, z=climatedata$p07)
 #Temperature is modeled without transformation, but precipitation should be log transformed before modeling, because variability in precipitation is proportional to total amount. Add positive value to monthly precipitation to avoid log zero error (this can be reversed after modeling, but may create negative values which should be set to zero).
 pts$z <- log10(pts$z + 1)
 #Establish cropped extent for northwest US.
 cropto <- c(-126,-105,35,50)
 #Run model with a parameter ensuring that coefficients for a tile are only estimated when neighborhood expanded to enough data points to include sufficient elevation range of 500 meters (this constraint applies only to the first covariate, which is elevation). If planning to incorporate into multiple layered raster, provide appropriate layer name.
 p07 <- toclimrast(pts, altlayer, cropto, covrange = 500, segx = 10, segy = 10); names(p07) <- 'p07'
 plot(p07, col=rev(map.pal('bcyr')))
 #Convert back to millimeters and show points
 p07mm <- 10^p07-1; p07mm <- ifel(p07mm<0,0,p07mm)
 plot(p07mm, breaks=c(0,5,10,20,50,100,200,400), col=rev(map.pal('bcyr')))
 vts <- vect(pts, geom=c("x", "y"),crs=crs('epsg:4326'))
 points(vts)

 #Temperatures
 climatedata <- read.csv(system.file("extdata", "pts.t.csv", package="climatools"))
 #Use actual station elevation as fourth column to calculate lapse rates instead of raster layer.
 pts <- data.frame(x=climatedata$lon, y=climatedata$lat, z=climatedata$t07, elev=climatedata$elev)
 #Crop to the Great Lakes
 cropto <- c(-95,-75, 40, 50)
 #When using raster to predict temperatures we do not need rain shadows, constrain covariates to elevation and proximity to water.
 t07 <- toclimrast(pts, altlayer[[1:2]], cropto, covrange = 200); names(t07) <- 't07'
 plot(t07, col=map.pal('bcyr'))
 #plot descrete intervals in Celsius.
 plot(t07, breaks=c(12:25), col=map.pal('bcyr'))
 vts <- vect(pts, geom=c("x", "y"),crs=crs('epsg:4326'))
 points(vts)
 #January temperature
 pts <- data.frame(x=climatedata$lon, y=climatedata$lat, z=climatedata$t01, elev=climatedata$elev)
 t01 <- toclimrast(pts, altlayer[[1:2]], cropto, covrange = 200); names(t01) <- 't01'
 #plot continuous
 plot(t01, col=map.pal('bcyr'))
 #Adding residuals via interpolation instead of random forest.
 t01a <- toclimrast(pts, altlayer[[1:2]], cropto, covrange = 200, randforest = FALSE); names(t01a) <- 't01'
 plot(t01a, col=map.pal('bcyr'))
 #add points to plot
 points(vts)
}
