% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{toclimrast}
\alias{toclimrast}
\title{Make climate raster from point data}
\usage{
toclimrast(
  pts,
  altlayer,
  cropto = NULL,
  covrange = 0,
  minrow = 50,
  segx = 5,
  segy = 5,
  cropbuffer = 5,
  randforest = TRUE
)
}
\arguments{
\item{pts}{data frame of 3 or 4 columns; first two: xy coordinates (longitude and latitude decimal degrees); third column is climate attribute like temperature; fourth column if provided is substitute values for first covariate (e.g. station elevation being more precise than gridded elevation).}

\item{altlayer}{terra raster layer (multiple layers) representing covariates. Elevation is recomended as first covariate if station elevation provided as fourth column in point data.}

\item{cropto}{Optional crop extent c(xmin, xmax, ymin, ymax).}

\item{covrange}{Minimum range in first covariate before building submodel coefficients.}

\item{minrow}{Minimum number of data points before building submodel coefficients.}

\item{segx}{Number of segments in x axis for building submodel coefficients.}

\item{segy}{Number of segments in y axis for building submodel coefficients.}

\item{cropbuffer}{Buffer around the crop extent to incorporate points into model outside cropped area (units of raster projection).}

\item{randforest}{Use random forest model if true (requires ranger package) to generate coefficients and residuals layers. Alternatively, coefficients and residuals layers are interpolated solely using xy coordinate using gstats package.}
}
\value{
Climate raster matching cropped extent.
}
\description{
Make climate raster from point data
}
