% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_tools.R
\name{tpi}
\alias{tpi}
\title{Topographic position index using 3 neighborhood scales}
\usage{
tpi(dm, r1, r2, r3)
}
\arguments{
\item{dm}{original raster}

\item{r1}{small neighborhood radius}

\item{r2}{intermediate neighborhood radius}

\item{r3}{large neighborhood radius}
}
\value{
Topographic position index raster
This function simply averages the hillslope positions for all 3 neighborhoods.
}
\description{
Topographic position index using 3 neighborhood scales
}
