% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XtremLow.R
\name{warm6ppt}
\alias{warm6ppt}
\title{Highest/Lowest n month period temperature/precipitation}
\usage{
warm6ppt(t, p, n = 6, max_var = TRUE, t_input = TRUE, p_return = TRUE)
}
\arguments{
\item{t}{vector of 12 monthly mean temperatures}

\item{p}{vector of 12 monthly mean precipitation}

\item{n}{number of months in period}

\item{max_var}{Use maximum if TRUE (minimum if FALSE)}

\item{t_input}{Assess period with extreme temperature if TRUE (extreme precipitation if FALSE)}

\item{p_return}{Return precipitation for the extreme period if TRUE (return temperature if FALSE)}
}
\value{
A single value for the temperature/precipitation of the warmest/coldest/wettest/driest n-month period.
}
\description{
Highest/Lowest n month period temperature/precipitation
}
\examples{
t=generateTemp(5,22)
p=generatePpt(100,50,150)
#precipitation of warmest 6 months
warm6ppt(t, p, n=6, max_var=TRUE, t_input=TRUE, p_return=TRUE)
#temperature of warmest 6 months
warm6ppt(t, p, n=6, max_var=TRUE, t_input=TRUE, p_return=FALSE)
#temperature of driest 3 months
warm6ppt(t, p, n=3, max_var=FALSE, t_input=FALSE, p_return=FALSE)
#precipitation of driest 3 months
warm6ppt(t, p, n=3, max_var=FALSE, t_input=FALSE, p_return=TRUE)
#temperature of wettest 3 months
warm6ppt(t, p, n=3, max_var=TRUE, t_input=FALSE, p_return=FALSE)
}
